fix(a2ui-v1-dialect): TASK-2537 — move basic_functions() into catalog/basic/__init__.py per task file list
